WPF-956968-Updated the sample and readme file
