+++ CustomLegendSample/App.xaml
@@ -1,6 +1,7 @@
-﻿<Application x:Class="Sample.App"
+﻿<Application x:Class="CustomLegendSample.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:local="clr-namespace:CustomLegendSample"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
