+++ CustomLegendSample/App.xaml.cs
@@ -1,12 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Configuration;
+﻿using System.Configuration;
 using System.Data;
-using System.Linq;
-using System.Threading.Tasks;
 using System.Windows;
 
-namespace Sample
+namespace CustomLegendSample
 {
     /// <summary>
     /// Interaction logic for App.xaml
@@ -14,4 +10,5 @@ namespace Sample
     public partial class App : Application
     {
     }
+
 }
