+++ CustomLegendSample/CustomLegendSample.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.12.35527.113 d17.12
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomLegendSample", "CustomLegendSample.csproj", "{3CCAEB5F-6EA0-45F3-8320-34B305EDE66D}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3CCAEB5F-6EA0-45F3-8320-34B305EDE66D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3CCAEB5F-6EA0-45F3-8320-34B305EDE66D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3CCAEB5F-6EA0-45F3-8320-34B305EDE66D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3CCAEB5F-6EA0-45F3-8320-34B305EDE66D}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
