+++ CustomLegendSample/MainWindow.xaml
@@ -1,44 +1,43 @@
-﻿<Window x:Class="Sample1.MainWindow"
+﻿<Window x:Class="CustomLegendSample.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:local="clr-namespace:Sample1"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:chart="clr-namespace:Syncfusion.UI.Xaml.Charts;assembly=Syncfusion.SfChart.WPF"
-        Title="Sample" WindowState="Maximized"
-        x:Name="window" 
-        Height="500" Width="725">
+        xmlns:local="clr-namespace:CustomLegendSample"
+        mc:Ignorable="d"
+        Title="Sample" WindowState="Maximized" x:Name="window" Height="500" Width="725">
+
     <Grid x:Name="grid" Margin="10">
         <Grid.DataContext>
-            <local:ViewModel x:Name="viewModel"></local:ViewModel>
+            <local:ViewModel x:Name="viewModel"/>
         </Grid.DataContext>
+
         <Grid.Resources>
             <DataTemplate x:Key="itemTemplate">
 
                 <StackPanel Margin="10,0,10,0"  Orientation="Horizontal">
 
                     <CheckBox Margin="2" Tag="{Binding}"
-                              IsChecked="True"
-                              Checked="CheckBox_Checked" Unchecked="CheckBox_Unchecked"/>
+                          IsChecked="True"
+                          Checked="CheckBox_Checked" Unchecked="CheckBox_Unchecked"/>
                     <Rectangle  Width="{Binding IconWidth}" Height="{Binding IconHeight}"
-                                        
-                            Fill="{Binding Interior}" Margin="2"/>
+                                    
+                        Fill="{Binding Interior}" Margin="2"/>
 
                     <TextBlock HorizontalAlignment="Center"
-                               Margin="5,0,0,0"
-                               VerticalAlignment="Center"
-                               Text="{Binding Item.XValue}">
+                           Margin="5,0,0,0"
+                           VerticalAlignment="Center"
+                           Text="{Binding Item.XValue}">
 
                     </TextBlock>
-
                 </StackPanel>
-
             </DataTemplate>
         </Grid.Resources>
-        <chart:SfChart   x:Name="chart">
-            
+
+        <chart:SfChart x:Name="chart">
             <local:PieSeriesExt x:Name="series1" Label="Series" XBindingPath="XValue" YBindingPath="YValue" />
             <local:PieSeriesExt x:Name="series2" XBindingPath="XValue" YBindingPath="YValue" />
-
         </chart:SfChart>
-       
-    </Grid>    
+    </Grid>
 </Window>
