+++ CustomLegendSample/Model/Model.cs
@@ -1,0 +1,8 @@
+﻿namespace CustomLegendSample
+{
+    public class Model
+    {
+        public string XValue { get; set; }
+        public double YValue { get; set; }
+    }
+}
