--- WPF_Sample/Sample.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26430.15
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sample", "Sample\Sample.csproj", "{A8AD51AF-004F-48DF-B8B5-79630D63844D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-		Release-Xml|Any CPU = Release-Xml|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A8AD51AF-004F-48DF-B8B5-79630D63844D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A8AD51AF-004F-48DF-B8B5-79630D63844D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A8AD51AF-004F-48DF-B8B5-79630D63844D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A8AD51AF-004F-48DF-B8B5-79630D63844D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A8AD51AF-004F-48DF-B8B5-79630D63844D}.Release-Xml|Any CPU.ActiveCfg = Release|Any CPU
-		{A8AD51AF-004F-48DF-B8B5-79630D63844D}.Release-Xml|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {F8FE655E-9E54-49B3-B229-EB811FF2AC0A}
-	EndGlobalSection
-EndGlobal
